fix: move all py files to scripts folder
